Revise times of polling mgmt port link status to 100
In rescue mode, ONIE configures network during booting only one time.
If the mgmt port doesn't link up, ONIE will not configure the port.

The polling loop of checking link status is a total 5 seconds in the
worst case. It may be timeout eventually if the link partner has
capibility of 1000 full duplex.

We ran test of booting on Accton AS5710_54x repeatedly and found it
has probability over 50 times polling. The test result are:

- 40 times (55%)
- 50 times (45%)
- 69 times (8%)
- 79 times (2%)