x86_64: vga console support
By default ONIE supports user I/O via a serial console connection, as
this is how all current physical hardware is built.  This comes into
play for:

- the GRUB configuration
- the Linux kernel configuration

This patch adds an option to support a 'regular' keyboard and console
connection, typical of a Linux workstation.

The Makefile option is SERIAL_CONSOLE_ENABLE with the default value of
"yes".

To use a regular keyboard and console connection, when building the
machine specify SERIAL_CONSOLE_ENABLE=no on the command line, for
example:

  make -j8 MACHINE=kvm_x86_64 all SERIAL_CONSOLE_ENABLE=no

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>