lvm2: default to version 2_02_105 for non-armv8a machines
The version of lvm2 was pushed ahead to 2_02_155 as part of the
aarch64 work.  It appears it was tested using glibc, but was never
tested using uClibc.

It is unclear why moving ahead to version 2_02_155 was necessary.
Looking at the upstream, version 2_02_155 was never officially
released.

This patch reverts the default version of lvm2 to 2_02_105 for all CPU
architectures except armv8a.  For armv8a machines, the default version
of lvm2 is left as 2_02_155.

Closes: #480
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>