build: compile common utilities once per toolchain
Today ONIE re-compiles *everything*, except the toolchain, for each
machine.  This is really wasteful of time, CPU and disk space when
building multiple machines that share a toolchain.

Most ONIE components are platform independent.  For example the common
utilities, like e2fstools, dropbear, etc., only need to be built once
for each toolchain.

Some components are machine specific and must be built for each
machine, such as:

- Linux kernel
- U-Boot
- busybox  (contains machine specific i2c EEPROM access)
- i2ctools (contains machine specific i2c EEPROM access)

This patch introduces the notion of common "user" components (short
for user space) that only need to be built once per toolchain.  This
significantly improves the build times when building multiple
machines.

With this change the build directory structure now contains a common
user build directory, called "build/user/<toolchain>".  All the common
components are built here.

The other change is the behavior of "make clean".  Historically "make
clean" cleaned up the build space for a single machine.  Now that
common component build products are shared across machine builds, it
would be bad to remove the common components as part of "make clean".

For cleaning, this patch introduces two new clean related targets:

- machine-clean -- only cleans components that are specific to the
  machine, like the kernel, u-boot, busybox, i2ctools, etc.

- user-clean -- a super set including machine-clean, plus all the
  common user components

The old "clean" target is now a synonym for "machine-clean".

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>